Conflicts:
	BattleBoy/BattleBoy.exe
	BattleBoy/BattleBoy.h
	BattleBoy/BattleBoy.pdb
	BattleBoy/CEGUI.log
	BattleBoy/DebugLog.txt
	Release/Boy.lib
	Release/Common.lib

+++ libs/Boy/Graphics.h
@@ -47,7 +47,7 @@ namespace Boy
 		 * geometric primitive drawing methods
 		 */
 		virtual void drawLine(float x0, float y0, float x1, float y1) = 0;
-		virtual void fillRect(int x0, int y0, int w, int h) = 0;
+		virtual void fillRect(float x0, float y0, float w, float h) = 0;
 		virtual void drawCircle(float x0, float y0, float r, int s) = 0;
 		virtual void drawTriStrip(TriStrip *strip) = 0;
 		virtual void drawLineStrip(TriStrip *strip) = 0;
